cart bug2 fix

+++ inventory/invention/views.py
@@ -27,6 +27,7 @@ from django.utils import timezone
 from rest_framework.response import Response
 from rest_framework.decorators import api_view, APIView
 from django.db import transaction
+from django_ratelimit.decorators import ratelimit
 
 
 #rest_api
@@ -67,6 +68,7 @@ def custom_forbidden(request):
 
 
 #Login-And-Register
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @unauthenticated_user
 def login(request):
     if request.user.is_authenticated:
@@ -87,6 +89,7 @@ def login(request):
  
     return render(request,'credential/login.html')
 
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @unauthenticated_user
 def signup(request):
     details=User.objects.all()   
@@ -113,6 +116,7 @@ def signup(request):
 
 
 #Home-Page
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @login_required(login_url= 'login')
 @allowed_user(allowed_roles=['student_user','admin', 'superadmin'])
 def home(request):
@@ -123,21 +127,30 @@ def home(request):
 
     products = Product.objects.all()
     print(products)
-    return render(request, 'core/home.html', {'products': products,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'core/home.html', {'products': products, 'count':count,})
 
 
 
 #View-Product-Details-As-View-Details
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @login_required(login_url='login')
 def product_description(request, pk):
     item = Product.objects.get(pk =pk)
-    return render(request, 'core/product_description.html', {'item':item,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'core/product_description.html', {'item':item, 'count':count,})
 
 
 #About-Page
 @login_required(login_url='login')
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 def about(request):
-	return render(request, 'core/about.html')
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    
+    return render(request, 'core/about.html', {'count':count,})
 
 
 #Access-Denied-Page
@@ -148,8 +161,8 @@ def no_permission(request):
 #Cart Functionality
 
 #Add-To-Cart
-temporary_cart = defaultdict(int)
-
+# temporary_cart = defaultdict(int)
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 def add_to_cart(request, product_id):
     product = Product.objects.select_for_update().get(id=product_id)
     if product.available_count == 0:
@@ -163,7 +176,7 @@ def add_to_cart(request, product_id):
                 quantity_int = int(quantity)
                 if 0 < quantity_int <= product.available_count and quantity_int <= product.dummy_count:
                     with transaction.atomic():
-                        if quantity_int <= product.dummy_count - temporary_cart[product_id]:
+                        if quantity_int <= product.dummy_count :
                             a=0
                             try:
                                 cart=Cart.objects.get(product_name=product,created_by=request.user)
@@ -172,17 +185,15 @@ def add_to_cart(request, product_id):
                                 a=cart.quantity
                                 a += quantity_int
                                 cart.quantity=a
+                                if a > product.dummy_count:
+                                    messages.warning(request, "Not Enough quantity available.")
+                                    return redirect('Home')
                                 cart=Cart.objects.get(product_name=product,created_by=request.user).delete()
-                                Cart.objects.create(product_name=product,quantity=a,created_by=request.user)
-                                
+                                Cart.objects.create(product_name=product,quantity=a,created_by=request.user)  
                             except:
                                 Cart.objects.create(product_name=product,quantity=quantity_int,created_by=request.user)
-                                pass
-                            
-                            # cart=Cart.objects.get(product_name=product,created_by=request.user).delete()
-                            temporary_cart[product_id] += quantity_int
                         else:
-                            messages.warning(request, "Not Enough quantity available.")
+                            messages.warning(request, "Not Enough quantity available...")
                 else:
                     messages.warning(request, "look up a valid quantity.")
             except ValueError:
@@ -194,7 +205,8 @@ def view_cart(request):
         cart=Cart.objects.filter(created_by=request.user)
         product=Product.objects.all()
         category=Category.objects.all()
-        return render(request, 'cart/cart.html', {'cart_items':cart,'product':product,'category':category})
+        count = cart.count()
+        return render(request, 'cart/cart.html', {'cart_items':cart,'product':product,'category':category, 'count':count,})
 
 #Remove-Cart
 def remove_from_cart(request, product_id):
@@ -204,32 +216,39 @@ def remove_from_cart(request, product_id):
 
 
 #Submit-In-Cart
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 def submit_cart(request):
     if request.method == "POST":
         due_date = request.POST.get('due_date')
         for i in Cart.objects.filter(created_by=request.user):
             status='checked_in'
-            PurchasedItem.objects.create(product=i.product_name, quantity=i.quantity, user=request.user,status=status,date_added=datetime.datetime.now(), due_date=due_date)
-            Log.objects.create(product=i.product_name, quantity=i.quantity, user=request.user, status=status, created_at=datetime.datetime.now(), due_date=due_date)
-            Cart.objects.filter(created_by=request.user).delete()
+            p=Product.objects.get(name=i.product_name)
+            if (p.available_count - i.quantity) < 0:
+                Cart.objects.filter(created_by=request.user).delete()
+                messages.warning(request, "Not Enough quantity available..")
+            else:
+                Product.objects.filter(name=i.product_name).update(available_count=F('available_count')-i.quantity)
+                PurchasedItem.objects.create(product=i.product_name, quantity=i.quantity, user=request.user,status=status,date_added=datetime.datetime.now(), due_date=due_date)
+                Log.objects.create(product=i.product_name, quantity=i.quantity, user=request.user, status=status, created_at=datetime.datetime.now(), due_date=due_date)
+                Cart.objects.filter(created_by=request.user).delete()
     return redirect("Home")
 
 
 #Optional-Cart-Quantity
-def decrease_quantity(request, product_id):
-    if product_id in temporary_cart:
-        temporary_cart[product_id] -= 1
-        if temporary_cart[product_id] == 0:
-            del temporary_cart[product_id]
-    return redirect('view_cart')
+# def decrease_quantity(request, product_id):
+#     if product_id in temporary_cart:
+#         temporary_cart[product_id] -= 1
+#         if temporary_cart[product_id] == 0:
+#             del temporary_cart[product_id]
+#     return redirect('view_cart')
 
 
-def increase_quantity(request, product_id):
-    if product_id in temporary_cart:
-        product = Product.objects.get(pk=product_id)
-        if product.available_count > temporary_cart[product_id]:
-            temporary_cart[product_id] += 1
-    return redirect('view_cart')
+# def increase_quantity(request, product_id):
+#     if product_id in temporary_cart:
+#         product = Product.objects.get(pk=product_id)
+#         if product.available_count > temporary_cart[product_id]:
+#             temporary_cart[product_id] += 1
+#     return redirect('view_cart')
 
 
 def update_quantity(request, product_id, quantity):
@@ -245,10 +264,13 @@ def update_quantity(request, product_id, quantity):
 
 #Return-Form-View
 def return_form(request):
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
     purchased_items = Log.objects.filter(user = request.user) 
-    return render(request, 'cart/return_form.html', {'purchased_items':purchased_items,})
+    return render(request, 'cart/return_form.html', {'purchased_items':purchased_items, 'count':count,})
 
 #Return-All
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 def return_all(request, item_id):
     item = get_object_or_404(Log, pk=item_id)
     product = item.product
@@ -270,10 +292,12 @@ class AddReturnView(View):
         categories = Category.objects.all()
         products = Log.objects.all()
         item = get_object_or_404(Log, id=item_id)
+        cart=Cart.objects.filter(created_by=request.user)
+        count = cart.count()
         return render(
             request,
             'cart/return.html',
-            {'categories': categories, 'products': products, 'item': item}
+            {'categories': categories, 'products': products, 'item': item, 'count':count}
         )
     
     def post(self, request, item_id):
@@ -301,10 +325,12 @@ class AddReturnView(View):
         else:
             pass
         
+        cart=Cart.objects.filter(created_by=request.user)
+        count = cart.count()
         return render(
             request,
             'cart/return.html',
-            {'categories': categories, 'products': products, 'item': item}
+            {'categories': categories, 'products': products, 'item': item, 'count':count,}
         )
 
 
@@ -315,10 +341,12 @@ class AddWastageView(View):
         categories = Category.objects.all()
         products = Log.objects.all()
         item = get_object_or_404(Log, id=item_id)
+        cart=Cart.objects.filter(created_by=request.user)
+        count = cart.count()
         return render(
             request,
             'cart/wastage.html',
-            {'categories': categories, 'products': products, 'item': item}
+            {'categories': categories, 'products': products, 'item': item, 'count':count,}
         )
 
     def post(self, request, item_id):
@@ -345,11 +373,12 @@ class AddWastageView(View):
             
         else:
             pass
-
+        cart=Cart.objects.filter(created_by=request.user)
+        count = cart.count()
         return render(
             request,
             'cart/wastage.html',
-            {'categories': categories, 'products': products, 'item': item}
+            {'categories': categories, 'products': products, 'item': item, 'count':count,}
         )
         
         
@@ -374,7 +403,9 @@ def users_list(request):
             AdminMail.objects.create(mail=email)
         users = User.objects.all()
         return redirect('users_list')
-    return render(request, 'superadmin_view/users.html', {'users': users,'admins':admins})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'superadmin_view/users.html', {'users': users,'admins':admins, 'count':count,})
 
 
 #Super-Admin-Remove-The-Admin-Role
@@ -405,18 +436,24 @@ def admin_view(request):
     log = Log.objects.all()
     purchased_items = PurchasedItem.objects.all()
     checked_out = CheckedOutLog.objects.all()
-    return render(request, 'adminview/admin.html', {'log':log, 'checked_out':checked_out, 'purchased_items': purchased_items,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'adminview/admin.html', {'log':log, 'checked_out':checked_out, 'purchased_items': purchased_items,'count':count,})
 
 
 #Wastage-Record-View-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=(['admin', 'superadmin']))
 @login_required(login_url='login')
 def wastage(request):
     wastage = Wastage.objects.all()
-    return render(request, 'adminview/wastage_render.html', {'wastage': wastage,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'adminview/wastage_render.html', {'wastage': wastage, 'count':count,})
 
 
 #Add-Product-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=(['admin', 'superadmin']))
 @login_required(login_url='login')
 def add_product(request):
@@ -438,18 +475,23 @@ def add_product(request):
          else:
              sweetify.warning(request, 'Product added successfully',button="OK")
              return redirect("Add_product")
-    
-   return render (request,"adminview/add_product.html",{"category":category, "products":products,})
+   cart=Cart.objects.filter(created_by=request.user)
+   count = cart.count()
+   return render (request,"adminview/add_product.html",{"category":category, "products":products,'count':count,})
 
 #View-Product-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def view_product(request):
     products = Product.objects.all()
-    return render(request, 'adminview/product.html', {'products':products,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'adminview/product.html', {'products':products, 'count':count,})
 
 
 #Remove-Product-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def remove_product(request, pk):
@@ -459,6 +501,7 @@ def remove_product(request, pk):
 
 
 #Add-Category-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def add_category(request):
@@ -467,18 +510,24 @@ def add_category(request):
      if request.method == "POST":
          name = request.POST.get('name')
          Category.objects.create(name = name, created_by = request.user)
-     return render(request, 'adminview/add_category.html', {'categories': categories,'existing_categories': list(existing_categories)})
+     cart=Cart.objects.filter(created_by=request.user)
+     count = cart.count()
+     return render(request, 'adminview/add_category.html', {'categories': categories,'existing_categories': list(existing_categories), 'count':count,})
 
 
 #View-Category-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def category(request):
         categories = Category.objects.all()
-        return render(request, 'adminview/editCategory.html', {'categories': categories,})
+        cart=Cart.objects.filter(created_by=request.user)
+        count = cart.count()
+        return render(request, 'adminview/editCategory.html', {'categories': categories, 'count':count,})
 
 
 #Remove-Category-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def remove_category(request, category_id):
@@ -488,6 +537,7 @@ def remove_category(request, category_id):
 
 
 #Edit-Category-For-Admin-SuperAdmin
+@ratelimit(key='ip', rate='10/m', method=ratelimit.ALL, block=True)
 @allowed_user(allowed_roles=['admin', 'superadmin'])
 @login_required(login_url='login')
 def edit_category(request, category_id):
@@ -499,6 +549,8 @@ def edit_category(request, category_id):
             category.name = new_category_name
             category.save()
             return redirect('Add_category')
-    return render(request, 'adminview/edit_category.html', {"category":category,})
+    cart=Cart.objects.filter(created_by=request.user)
+    count = cart.count()
+    return render(request, 'adminview/edit_category.html', {"category":category, 'count':count,})
